04/requirements.txt
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.8.2
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
cached-property==1.5.1
ccsm==0.9.14.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
curl==0.0.1
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
dnspython==2.0.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
grab==0.6.41
grpcio==1.16.1
Guake==3.6.3
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
importlib-metadata==1.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.5.2
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==18.0.1
kiwisolver==1.2.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
lxml==4.5.2
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.2
mccabe==0.6.1
memory-profiler==0.58.0
mistune==0.8.4
more-itertools==4.2.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
nose==1.3.7
notebook==6.1.5
numpy==1.19.2
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.1
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip-search==0.0.6
plotext==2.2.2
prettytable==2.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.6.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.6
Pygments==2.7.2
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-nmap==0.6.1
python-xapp==2.0.1
python-xlib==0.23
pythonping==1.0.15
pytils==0.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selection==0.0.14
selenium==3.141.0
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.15.0
soupsieve==1.9.5
systemd-python==234
tabulate==0.8.9
terminado==0.9.1
terminalplot==0.3.0
testpath==0.4.4
texttable==1.6.2
toml==0.10.1
tornado==5.1.1
traitlets==4.3.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
user-agent==0.1.9
vboxapi==1.0
virtualenv==20.1.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
weblib==0.1.30
websocket-client==0.53.0
wrapt==1.12.1
xkit==0.0.0
zeroconf==0.24.4
zipp==1.0.0